back/requirements.txt
appdirs==1.4.3
decorator==4.1.2
Django==1.11.4
django-easy-select2==1.3.4
graphene==1.4.1
graphene-django==1.3
graphql-core==1.1
graphql-relay==0.4.5
gunicorn==19.7.1
ipython==6.1.0
ipython-genutils==0.2.0
iso8601==0.1.12
jedi==0.10.2
packaging==16.8
pexpect==4.2.1
pickleshare==0.7.4
promise==2.0.2
prompt-toolkit==1.0.15
ptyprocess==0.5.2
py==1.4.34
Pygments==2.2.0
pyparsing==2.2.0
pytest==3.2.1
pytest-django==3.1.2
pytz==2017.2
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
traitlets==4.3.2
typing==3.6.2
wcwidth==0.1.7
whitenoise==3.3.0
